+++ src/write/error.rs
@@ -4,7 +4,6 @@ use std::task::Poll;
 
 #[derive(Debug)]
 pub enum WriteError {
-    UnexpectedEof { expected: u64, received: u64 },
     WriteZero,
     OverlappingEntry,
 }
@@ -13,7 +12,6 @@ impl WriteError {
     #[inline]
     pub fn kind(&self) -> ErrorKind {
         match self {
-            Self::UnexpectedEof { .. } => ErrorKind::UnexpectedEof,
             Self::WriteZero => ErrorKind::WriteZero,
             Self::OverlappingEntry => ErrorKind::Unsupported,
         }
@@ -25,10 +23,6 @@ impl std::error::Error for WriteError {}
 impl fmt::Display for WriteError {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::UnexpectedEof { expected, received } => format!(
-                "expecting more data for entry; expected = {expected}, received = {received}"
-            )
-            .fmt(f),
             Self::WriteZero => "failed to write the buffered data".fmt(f),
             Self::OverlappingEntry => {
                 "cannot write new entry while another is being written".fmt(f)